refactor resources_test mocks to be included in mocks
